Rebase Rumble to head

+++ src/BizHawk.Bizware.OpenTK3/OTK_Gamepad.cs
@@ -367,7 +367,7 @@ namespace BizHawk.Bizware.OpenTK3
 		/// <remarks><paramref name="left"/> and <paramref name="right"/> are in 0..<see cref="int.MaxValue"/></remarks>
 		public void SetVibration(int left, int right)
 		{
-			static float Conv(int i) => (float)i/int.MaxValue;
+			static float Conv(int i) => (float)((double)i/int.MaxValue); // Preserve some precision
 			OpenTKGamePad.SetVibration(_deviceIndex, Conv(left), Conv(right));
 		}
 	}
